Add printf levels.  That way, vfprintf() can be conditionally compiled into more than one object file, with different space requirements and feature set.
The default (level 2) is to compile a version that offers integer and
string formatting, plus any useful modifiers (alignment, field width,
precision) applicable to it.  This is the version that goes into the
base libc.a.

The `minimal' version (level 1) leaves all modifier functionality out
except for the `l' modifier (argument is long int).  This will cause a
saving of several hundred bytes in flash consumption.  This version
goes into a separate libprintf_min.a that needs to be explicitly
requested when linking (-lprintf_min).  Note that in order to actually
get the minimal version linked, it needs to be referenced before.  This
can be forced using -Wl,-u,vfprintf.

An additional level 3 is planned that implements floating point
conversions as well.

FIXME: documentation not yet updated.


git-svn-id: http://svn.savannah.nongnu.org/svn/avr-libc/trunk@143 1084037a-409f-4a99-8160-1a4b3577d4c1