Allow the twitest example to run off the default 1 MHz RC oscillator.
This requires using the U2X bit in the UART initialization in order to
get an acceptable baud rate error for a 9600 Bd communication, as well
as a slowdown for the maximal TWI clock rate in order to maintain the
minimum required TWBR value per the datasheet.

While I was at it, fixed a couple of whitespace errors.

The new annotations required to renumber almost all annotations.


git-svn-id: http://svn.savannah.nongnu.org/svn/avr-libc/trunk@559 1084037a-409f-4a99-8160-1a4b3577d4c1