Provide a solution to the ``EEPROM library not working for a growing number of devices'' bug that gets away without a full per-device lib approach.
The solution is to build one set of EEPROM functions for each class of
devices with a known group of EEPROM registers.  The correct function
is then automatically selected based on a group suffix that is
silently added by the wrappers in <avr/eeprom.h>.

Thanks and kudos to Björn Haase for providing a turn-key solution for
this problem, as well as for anybody from the avr-libc developers list
who reviewed Björn's patch, so it would subsequently refined into a
ready-to-go solution.


git-svn-id: http://svn.savannah.nongnu.org/svn/avr-libc/trunk/avr-libc@785 1084037a-409f-4a99-8160-1a4b3577d4c1